Refactor configuration handling
This change should make it a bit easier to see when configuration is
read in, and would allow to further centralize parsing and loading of
the conf file.

We should not 'import config' each time when we need to get access to
the settings. Instead, the settings should be loaded once, and then
explicitly passed around. This patch does not make that happen yet, but
is the first step.

Change-Id: If9ad1222ceb806337a745845ee6e76605bab5fd3